requirements.txt
svgpathtools
geojson
shapely
tqdm
python-dotenv
psycopg2-binary